% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_summaries_helper_functions.R
\name{annual_rainfall_annual_rain}
\alias{annual_rainfall_annual_rain}
\title{Calculate Annual Rainfall}
\usage{
annual_rainfall_annual_rain(definitions, daily, data_names)
}
\arguments{
\item{definitions}{A list containing definitions for the annual rainfall analysis,
including options for calculating total rainfall, number of rainy days,
and criteria for handling missing values.}

\item{daily}{A data frame containing daily rainfall data.}

\item{data_names}{A list of variable names used in the data frame.}
}
\value{
A numeric vector representing annual rainfall and related calculations.
}
\description{
This function calculates annual rainfall based on the provided definitions,
daily rainfall data, and data variable names.
}
\examples{

# # Example usage:
# require(dplyr)
# library(cdms.products)
# data(daily_niger)
# definitions <- list(
#   annual_rain = list(
#     annual_rain = TRUE,
#     n_rain = FALSE,
#     na_rm = TRUE
#   )
# )
# data_names <- list(date = "date", station = "station_name", year = "year", rain = "rain")
# daily_data <- daily_niger \%>\% dplyr::filter(year > 1975) \%>\% dplyr::filter(station_name == "Zinder")
# annual_rainfall <- annual_rainfall_annual_rain(definitions, daily_data, data_names)

}
