% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_rainfall_end_rains.R
\name{annual_rainfall_end_rains}
\alias{annual_rainfall_end_rains}
\title{Calculate End of Rains}
\usage{
annual_rainfall_end_rains(definitions, daily, data_names)
}
\arguments{
\item{definitions}{A list containing definitions for the end of the rainy season analysis,
including start and end days, interval length, and minimum rainfall criteria.}

\item{daily}{A data frame containing daily rainfall data.}

\item{data_names}{A list of variable names used in the data frame.}
}
\value{
A numeric vector representing the end of the rainy season in terms of day of the year (DOY).
}
\description{
This function calculates the end of the rainy season based on the provided definitions,
daily rainfall data, and data variable names.
}
\examples{

# Example usage:
require(dplyr)
data(daily_niger)
definitions <- list(
  end_rains = list(
    start_day = 122,
    end_day = 366,
    interval_length = 7,
    min_rainfall = 5
  )
)
data_names <- list(date = "date", station = "station_name", year = "year", rain = "rain")
daily_data <- daily_niger \%>\% dplyr::filter(year > 1975) \%>\% dplyr::filter(station_name == "Zinder")
end_of_rains <- annual_rainfall_end_rains(definitions, daily_data, data_names)

}
