% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_rainfall_end_season.R
\name{annual_rainfall_end_season}
\alias{annual_rainfall_end_season}
\title{Calculate End of Season}
\usage{
annual_rainfall_end_season(definitions, daily, data_names)
}
\arguments{
\item{definitions}{A list containing definitions for the end of the rainy season analysis,
including start and end days, capacity, water balance, evaporation criteria,
and related parameters.}

\item{daily}{A data frame containing daily rainfall data.}

\item{data_names}{A list of variable names used in the data frame.}
}
\value{
A numeric vector representing the end of the rainy season in terms of day of the year (DOY).
}
\description{
This function calculates the end of the rainy season based on the provided definitions,
daily rainfall data, and data variable names.
}
\examples{
# Example usage:
# require(dplyr)
# library(cdms.products)
# data(daily_niger)
# definitions <- list(
#   end_season = list(
#     start_day = 150,
#     end_day = 366,
#     capacity = 100,
#     water_balance_max = 50,
#     evaporation = "value",
#     evaporation_value = 10
#   )
# )
# data_names <- list(date = "date", station = "station_name", year = "year", rain = "rain")
# daily_data <- daily_niger \%>\% dplyr::filter(year > 1975) \%>\% dplyr::filter(station_name == "Zinder")
# end_of_season <- annual_rainfall_end_season(definitions, daily_data, data_names)

}
