% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_summaries_helper_functions.R
\name{annual_rainfall_seasonal_length}
\alias{annual_rainfall_seasonal_length}
\title{Calculate Seasonal Length}
\usage{
annual_rainfall_seasonal_length(
  definitions,
  daily,
  summary_data,
  data_names,
  summaries
)
}
\arguments{
\item{definitions}{A list containing definitions for various aspects of the analysis.}

\item{daily}{A data frame containing daily rainfall data.}

\item{summary_data}{A data frame containing summary data.}

\item{data_names}{A list of variable names used in the data frames.}

\item{summaries}{A character vector specifying the summary types.}
}
\value{
A numeric vector representing the seasonal length of annual rainfall.
}
\description{
This function calculates the seasonal length of annual rainfall based on the
provided definitions, daily rainfall data, summary data, data variable names,
and summary types.
}
\examples{

# Example usage:
definitions <- list(seasonal_length = list(end_type = "rains"))
data_names <- list(date = "date", station = "station", year = "year", rain = "rain")
summary_types <- c("start_rains", "end_rains", "end_season")
daily_data <- data.frame(date = seq(as.Date("2023-01-01"), as.Date("2023-12-31"), by = "days"),
                         station = "Station_A",
                         year = rep(2023, each = 365),
                         rain = runif(365, min = 0, max = 10))
summary_data <- data.frame(year = 2023, start_rains_date = as.Date("2023-04-01"), 
                           end_rains_date = as.Date("2023-10-31"), end_season_date = as.Date("2023-12-31"))
#seasonal_length <- annual_rainfall_seasonal_length(definitions, daily_data, summary_data, data_names, summary_types)
}
