% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_rainfall_summaries.R
\name{annual_rainfall_summaries}
\alias{annual_rainfall_summaries}
\title{Annual Rainfall Summaries}
\usage{
annual_rainfall_summaries(
  station_id,
  summaries = c("seasonal_rainfall", "seasonal_raindays", "start_rains", "end_season",
    "length_season")
)
}
\arguments{
\item{station_id}{\code{character} The id's of the stations to analyse. Either a
single value or a vector.}

\item{summaries}{\code{character} The names of the summaries to produce.}
}
\value{
A data frame with yearly summaries.
}
\description{
Annual Rainfall Summaries
}
\examples{
annual_rainfall_summaries(station_id = "1357")
}
