% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_annual_temperature.R,
%   R/update_monthly_temperature.R, R/update_seasonal_rain.R
\name{as_logical}
\alias{as_logical}
\title{Safely convert an Object to a Logical value}
\usage{
as_logical(x)

as_logical(x)

as_logical(x)
}
\arguments{
\item{x}{An object to be converted to a logical value If \code{x} is \code{NULL}, the function returns \code{NULL} instead of an error.}
}
\value{
A logical value if conversion is successful, or \code{NULL} if the input is \code{NULL}.

A logical value if conversion is successful, or \code{NULL} if the input is \code{NULL}.

A logical value if conversion is successful, or \code{NULL} if the input is \code{NULL}.
}
\description{
Safely convert an Object to a Logical value

Safely convert an Object to a Logical value

Safely convert an Object to a Logical value
}
