% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_seasonal_rain.R, R/update_start_rains.R
\name{as_numeric}
\alias{as_numeric}
\title{Safely convert an Object to Numeric}
\usage{
as_numeric(x)

as_numeric(x)
}
\arguments{
\item{x}{An object to be converted to numeric. If \code{x} is \code{NULL}, the function returns \code{NULL} instead of an error.}
}
\value{
A numeric value if conversion is successful, or \code{NULL} if the input is \code{NULL}.

A numeric value if conversion is successful, or \code{NULL} if the input is \code{NULL}.
}
\description{
Safely convert an Object to Numeric

Safely convert an Object to Numeric
}
