% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_crop_definitions.R
\name{build_crop_definitions}
\alias{build_crop_definitions}
\title{Build Crop Definitions from File}
\usage{
build_crop_definitions(definition_file = NULL)
}
\arguments{
\item{definition_file}{A list containing file data and attributes generated in R-Instat with
named vectors \code{Var1}, \code{Var2}, and \code{Var3} for water requirements, planting dates, and planting length respectively.}
}
\value{
A list representing the structured crop definition data, including water requirements,
planting dates, and planting length.
}
\description{
This function reads crop definition data from a provided file structure generated in R-Instat.
It then extracts information about water requirements, planting dates, and planting length for different crops.
The extracted values are then split into lists..
}
\examples{
# Assuming definition_file is a correctly structured list:
#get_crop_definitions(definition_file)
}
