% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_season_start_probabilities.R
\name{build_season_start_probabilities}
\alias{build_season_start_probabilities}
\title{Build Season Start Probabilities from File}
\usage{
build_season_start_probabilities(definition_file = NULL)
}
\arguments{
\item{definition_file}{A list containing file data and attributes generated in R-Instat.}
}
\value{
A list representing the season start probabilities with the specified days.
}
\description{
This function processes a file structure to extract information about the specified day for season start probabilities.
The information is split into lists
}
\examples{
#get_season_start_probabilities(definition_file)
}
