% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_definitions.R
\name{build_spells_summaries}
\alias{build_spells_summaries}
\title{Build Spells Definitions from File}
\usage{
build_spells_summaries(spells_data = NULL)
}
\arguments{
\item{spells_data}{Data frame used for spells data (usually \code{spells}).}
}
\value{
A list representing the structured crop definition data, including water requirements,
planting dates, and planting length.
}
\description{
This function reads spell data from a provided file structure generated in R-Instat.
It then extracts information about the parameter values given to construct these spells.
}
\examples{
# Assuming definition_file is a correctly structured list:
#get_crop_definitions(definition_file)
}
