% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_definitions_helper_functions.R
\name{create_end_rains_definitions}
\alias{create_end_rains_definitions}
\title{Get end rains definitions}
\usage{
create_end_rains_definitions(end_rains = NULL)
}
\arguments{
\item{end_rains}{The end rains data.}
}
\value{
A list representation of end rains definitions.
}
\description{
Retrieves end rains definitions.
}
\examples{
# Example usage:
#create_end_rains_definitions(end_rains)
}
