% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_definitions_helper_functions.R
\name{create_end_season_definitions}
\alias{create_end_season_definitions}
\title{Get end of season definitions}
\usage{
create_end_season_definitions(end_season = NULL)
}
\arguments{
\item{end_season}{The end season data.}
}
\value{
A list representation of end season definitions.
}
\description{
Retrieves end season definitions.
}
\examples{
# Example usage:
#create_end_season_definitions(end_season)
}
