% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_amendments.R
\name{date_amendments}
\alias{date_amendments}
\title{Date Amendments}
\usage{
date_amendments(summary, data, variable, definitions)
}
\arguments{
\item{summary}{A character string specifying the summary type.}

\item{data}{A data frame containing the original data.}

\item{variable}{A character string specifying the variable name in the data frame.}

\item{definitions}{A list containing definitions for summary types.}
}
\value{
The modified data frame with adjusted dates and day of year values.
}
\description{
This function performs date amendments based on the provided parameters.
It adjusts the year, day of year (DOY), and calculates new date values.
}
\examples{
summary <- "example_summary"
data <- data.frame(year = c(2023, 2024, 2025), variable = c(1, 200, 366))
variable <- "variable"
definitions <- list(example_summary = list(s_start_doy = 1))
date_amendments(summary, data, variable, definitions)
if shift year, then do this

}
