% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\title{Definitions}
\usage{
definitions(country, station_id, summaries, file = NULL)
}
\arguments{
\item{country}{\code{character(1)} The country code of the data.}

\item{station_id}{\code{character(1)} The station code in the data.}

\item{summaries}{\code{character} Vector of summaries to display}
}
\value{
TODO
}
\description{
Definitions
}
\examples{
# e.g. definitions("zm", "16", "annual_rain")
# error: definitions("zm", "1", c("annual_rain", "hi", "end_season"))
}
