% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_r_instat_to_bucket.R
\name{export_r_instat_to_bucket}
\alias{export_r_instat_to_bucket}
\title{Export R-Instat Data to Google Cloud Storage Bucket}
\usage{
export_r_instat_to_bucket(
  data,
  data_by_year,
  data_by_year_month = NULL,
  crop_data_name = NULL,
  rain = NULL,
  year = NULL,
  month = NULL,
  summaries = c("annual_rainfall", "annual_temperature", "monthly_temperature",
    "extremes", "crop_success", "start_season"),
  station_id,
  definitions_id,
  country,
  include_summary_data = FALSE,
  annual_rainfall_data = NULL,
  annual_temperature_data = NULL,
  monthly_temperature_data = NULL,
  crop_success_data = NULL,
  season_start_data = NULL,
  start_rains_column = "start_rain",
  start_rains_status_column = "start_rain_status",
  end_rains_column = "end_rains",
  end_rains_status_column = "end_rains_status",
  end_season_column = "end_season",
  end_season_status_column = "end_season_status",
  seasonal_length_column = "seasonal_length",
  min_tmin_column = "min_tmin",
  mean_tmin_column = "mean_tmin",
  max_tmin_column = "max_tmin",
  min_tmax_column = "min_tmax",
  mean_tmax_column = "mean_tmax",
  max_tmax_column = "max_tmax"
)
}
\arguments{
\item{data}{The main dataset.}

\item{data_by_year}{The dataset grouped by year.}

\item{data_by_year_month}{The dataset grouped by year and month.}

\item{crop_data_name}{Name of the crop data used for definitions when \code{summaries = "crop_success"}.}

\item{rain}{The rainfall data.}

\item{year}{The year data.}

\item{month}{The month data.}

\item{summaries}{A character vector specifying the types of summaries to include.}

\item{station_id}{\code{character} The id's of the stations to analyse. Either a
single value or a vector.}

\item{definitions_id}{\code{character} The ID to give to the definition file.}

\item{country}{\code{character(1)} The country code of the data.}

\item{include_summary_data}{Logical indicating whether to include summary data in the export.}

\item{annual_rainfall_data}{Annual rainfall summary data.}

\item{annual_temperature_data}{Annual temperature summary data.}

\item{monthly_temperature_data}{Monthly temperature summary data.}

\item{crop_success_data}{The proportion crop data output. Used if \code{summaries = "crop_success"} when \code{include_summary_data = TRUE}. This is called \code{crop_prop} by default in R-Instat.}

\item{season_start_data}{The crop data to be read into \code{"season_start_probabilities"}. This is called \code{crop_def} by default in R-Instat.}

\item{start_rains_column}{The name of the start of rains column in the data.}

\item{start_rains_status_column}{The name of the start of rains status column in the data.}

\item{end_rains_column}{The name of the end of rains column in the data.}

\item{end_rains_status_column}{The name of the end of rains status column in the data.}

\item{end_season_column}{The name of the end of season column in the data.}

\item{end_season_status_column}{The name of the end of season status column in the data.}

\item{seasonal_length_column}{The name of the seasonal length column in the data.}

\item{min_tmin_column}{The name of the minimum of minimum temperature column in the data.}

\item{mean_tmin_column}{The name of the mean of minimum temperature column in the data.}

\item{max_tmin_column}{The name of the maximum of minimum temperature column in the data.}

\item{min_tmax_column}{The name of the minimum of maximum temperature column in the data.}

\item{mean_tmax_column}{The name of the mean of maximum temperature column in the data.}

\item{max_tmax_column}{The name of the maximum of maximum temperature column in the data.}
}
\value{
A message confirming that the data has been uploaded to the bucket.
}
\description{
This function exports R-Instat data to a specified Google Cloud Storage bucket.
It collates the data and saves it as a JSON file on the local machine, then uploads
the file to the specified bucket.
}
\details{
This function collates the specified data into a JSON format and saves it to the local machine.
Then it uploads the JSON file to the specified Google Cloud Storage bucket. If \code{include_summary_data} is TRUE,
it also uploads additional summary data to the bucket.
}
\examples{
# Provide examples here if needed
}
