% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_most_recent_json.R
\name{extract_most_recent_json}
\alias{extract_most_recent_json}
\title{Extract the most recent JSON file from a list of filenames}
\usage{
extract_most_recent_json(files)
}
\arguments{
\item{files}{A character vector containing filenames.}
}
\value{
A character string representing the filename of the most recent JSON file.
}
\description{
This function takes a list of filenames as input and returns the filename
of the most recent JSON file. The filenames should be in the format
"definitions/1.YYYYMMDDHHMMSS.json".
}
\examples{
files <- c(
  "definitions/1.20240116155433.json",
  "definitions/1.20240304125111.json",
  "definitions/1.json"
)
epicsadata:::extract_most_recent_json(files)
}
