% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes_tmax.R
\name{extremes_tmax}
\alias{extremes_tmax}
\title{Get the Extreme Data for Maximum Temperature}
\usage{
extremes_tmax(country, station_id)
}
\arguments{
\item{country}{A character string specifying the country code of the data.}

\item{station_id}{A character vector specifying the ID(s) of the station(s) to analyse.}
}
\value{
A data frame containing the extreme data.
}
\description{
This function identifies extreme values for naximum temperature in a data frame.
}
\examples{
# Generate annual temperature summaries for station 16 in Zambia
#extremes_rain(country, station_id)
}
