% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_auth_file.R
\name{gcs_auth_file}
\alias{gcs_auth_file}
\title{GCS Auth with json file}
\usage{
gcs_auth_file(filename)
}
\arguments{
\item{filename}{The path to the JSON key file for GCS authentication.}
}
\value{
None
}
\description{
This function authenticates the user for Google Cloud Storage (GCS) using a JSON key file.
}
\examples{
# gcs_auth_file("path/to/key.json")

}
\references{
For more information on GCS authentication, refer to the official documentation:
https://cloud.google.com/storage/docs/reference/libraries#client-libraries-usage-r
}
