% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climsoft_connection_functions.R
\name{get_climsoft_conn}
\alias{get_climsoft_conn}
\title{Get Climsoft Connection}
\usage{
get_climsoft_conn()
}
\value{
The database connection object.
}
\description{
Retrieves the stored Climsoft database connection from the package environment.
}
\examples{
#con <- get_climsoft_conn()

}
