% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_daily_data}
\alias{get_daily_data}
\title{Get Daily Data}
\usage{
get_daily_data(country, station_id, call_from = c("climsoft", "googlebuckets"))
}
\arguments{
\item{country}{A character vector specifying the country or countries from which to get the data. Common options are \code{"mz"}, \code{"zm"}, and \code{"zm_test"}. Any defined in \code{get_bucket_name()}.}

\item{station_id}{A character string specifying the ID of the station for which to get the daily data.}

\item{call_from}{A character string specifying the location of the raw data.}
}
\value{
A data frame containing the daily data for the specified station and country.
}
\description{
Get Daily Data
}
\examples{
#
}
