% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from Google Cloud Storage}
\usage{
get_data(country, filename, save_to = NULL)
}
\arguments{
\item{country}{A character string specifying the country.}

\item{filename}{A character string specifying the name of the file in Google Cloud Storage.}

\item{save_to}{(Optional) A character string specifying the local path where the file should be saved. If not provided, the data will be parsed and returned as an R object.}
}
\value{
The function returns the retrieved data as an R object if \code{save_to} is \code{NULL.} If \code{save_to} is provided, the function saves the data locally.
}
\description{
This function retrieves data from Google Cloud Storage for a specified country and file. The data can either be parsed and returned as an R object or saved to a local disk.
}
\details{
The \code{get_data} function retrieves data from a specified country and file stored in Google Cloud Storage. It uses the \code{googleCloudStorageR} package
If the \code{save_to} argument is not provided or set to \code{NULL}, the function will parse the data and return it as an R object using the \code{googleCloudStorageR::gcs_parse_rds} function. This is useful when you want to directly work with the data in R, and can be more efficient.
If the \code{save_to} argument is provided with a valid local path, the function will save the file to the specified location on the disk using the \code{saveToDisk} parameter. This is useful when you want to download the file for further processing or analysis outside of R.
}
\examples{
# TODO
}
