% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_summary_functions.R
\name{get_definitions_id_from_metadata}
\alias{get_definitions_id_from_metadata}
\title{Get Definitions ID from Metadata}
\usage{
get_definitions_id_from_metadata(country, station_id)
}
\arguments{
\item{country}{A character string representing the country code.}

\item{station_id}{A character string representing the station ID.}
}
\value{
A character string representing the definitions ID from the station metadata.
}
\description{
This function retrieves the definitions ID from station metadata for a given country and station ID.
}
