% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_end_rains_definitions.R
\name{get_end_rains_definitions}
\alias{get_end_rains_definitions}
\title{Get end rains definitions}
\usage{
get_end_rains_definitions(end_rains = NULL)
}
\arguments{
\item{end_rains}{The end rains data.}
}
\value{
A list representation of end rains definitions.
}
\description{
Retrieves end rains definitions.
}
\examples{
# Example usage:
#get_end_rains_definitions(end_rains)
}
