% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast_data.R
\name{get_forecast_data}
\alias{get_forecast_data}
\title{Get Forecast Data}
\usage{
get_forecast_data(country, station_id)
}
\arguments{
\item{country}{A character vector specifying the country or countries from which to get the forecast data. Options are defined in \code{get_bucket_name()}.}

\item{station_id}{A character string specifying the ID of the station for which to get the forecast data.}
}
\value{
todo
}
\description{
Get Forecast Data
}
\examples{
# todo
}
