% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objects_in_bucket.R
\name{get_objects_in_bucket}
\alias{get_objects_in_bucket}
\title{Import Summary Definitions}
\usage{
get_objects_in_bucket(country, station_id, timestamp)
}
\arguments{
\item{country}{A character string specifying the country.}

\item{station_id}{A character string specifying the station ID.}

\item{timestamp}{The timestamp on the object file name to import. Default \code{NULL}}
}
\value{
A list containing imported or generated summary definitions.
}
\description{
This function imports summary definitions based on the country, station ID, summaries, and get_summaries.
It checks if there are any files corresponding to the provided station and summary in the Google Cloud Storage bucket.
If files are found, it imports definitions from the file; otherwise, it generates new definitions.
}
\examples{
# Import summary definitions
#import_summary_definitions("USA", "station123", list("info1", "info2"))

}
