% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_offset_term.R
\name{get_offset_term}
\alias{get_offset_term}
\title{Get Offset Term}
\usage{
get_offset_term(data)
}
\arguments{
\item{data}{A data object from which to retrieve the start DOY. This data object is expected
to have an associated metadata containing a \code{doy_start} field.}
}
\value{
Returns a single start DOY value if found in the metadata; otherwise, returns \code{NULL}.
}
\description{
This function retrieves the start day of the year (DOY) from the metadata of the given data.
If there are multiple start DOYs, it issues a warning and selects the first one.
}
\examples{
#data <- some_function_to_get_data()
#offset_term <- get_offset_term(data)
#print(offset_term)
}
