% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_bucket_helper_functions.R
\name{get_r_instat_definitions}
\alias{get_r_instat_definitions}
\title{Get R-Instat definitions}
\usage{
get_r_instat_definitions(calculation)
}
\arguments{
\item{calculation}{A list of calculations.}
}
\value{
A list of R-Instat definitions.
}
\description{
Retrieves R-Instat definitions based on given calculations.
}
\examples{
# Example usage:
#get_r_instat_definitions()
}
