% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_definitions_helper_functions.R
\name{get_season_length_definitions}
\alias{get_season_length_definitions}
\title{Get season length definitions}
\usage{
get_season_length_definitions(length = NULL)
}
\arguments{
\item{length}{The season length data.}
}
\value{
A list representation of season length definitions.
}
\description{
Retrieves season length definitions.
}
\examples{
# Example usage:
#get_season_length_definitions(length)
}
