% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_rains_definitions.R
\name{get_start_rains_definitions}
\alias{get_start_rains_definitions}
\title{Get start of rains definitions}
\usage{
get_start_rains_definitions(start_rains = NULL)
}
\arguments{
\item{start_rains}{The start rains data.}
}
\value{
A list of start of rains definitions
}
\description{
Retrieves start rains definitions.
}
\examples{
# Example usage:
#get_start_rains_definitions(start_rains)
}
