% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summaries_data.R
\name{get_summaries_data}
\alias{get_summaries_data}
\title{Get Summaries Data}
\usage{
get_summaries_data(country, station_id, summary)
}
\arguments{
\item{country}{A character vector specifying the country or countries from which to get the data. Options are defined in \code{get_bucket_name()} (e.g., \code{"zm"}, \code{"mw"}).}

\item{station_id}{A character string specifying the ID of the station for which to get the summary data.}

\item{summary}{A character string specifying the summary to retrieve.}
}
\value{
A list of data frames containing the summary data for the specified stations and country.
}
\description{
Get Summaries Data
}
\examples{
#

}
