% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp_summaries.R
\name{get_temp_summaries}
\alias{get_temp_summaries}
\title{Get temperature summaries}
\usage{
get_temp_summaries(temp_summary_name, data, to = c("annual", "monthly"))
}
\arguments{
\item{temp_summary_name}{Character vector specifying the name of the temperature summary.}

\item{data}{A list of temperature summaries by definition (e.g., year or month).}

\item{to}{A vector stating if it is annual or monthly summaries.}
}
\value{
A list containing temperature summary information.
}
\description{
Retrieves temperature summaries based on provided parameters.
}
\examples{
# Example usage:
}
