% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temperature_summary_definitions.R
\name{get_temperature_summary_definitions}
\alias{get_temperature_summary_definitions}
\title{Get temperature summary definitions}
\usage{
get_temperature_summary_definitions(
  year = data_book$get_climatic_column_name(data_name, "year"),
  month = data_book$get_climatic_column_name(data_name, "month"),
  data_by_year,
  data_by_year_month = NULL,
  min_tmin_column,
  mean_tmin_column,
  max_tmin_column,
  min_tmax_column,
  mean_tmax_column,
  max_tmax_column
)
}
\arguments{
\item{year}{Character vector specifying the year.}

\item{month}{Character vector specifying the month.}

\item{data_by_year}{A list of temperature summaries by definition (e.g., year).}

\item{data_by_year_month}{An optional second list of temperature summaries by definition (e.g., year and month).}

\item{min_tmin_column}{The name of the minimum of minimum temperature column in the data.}

\item{mean_tmin_column}{The name of the mean of minimum temperature column in the data.}

\item{max_tmin_column}{The name of the maximum of minimum temperature column in the data.}

\item{min_tmax_column}{The name of the minimum of maximum temperature column in the data.}

\item{mean_tmax_column}{The name of the mean of maximum temperature column in the data.}

\item{max_tmax_column}{The name of the maximum of maximum temperature column in the data.}
}
\value{
A list containing temperature summary definitions.
}
\description{
Retrieves temperature summary definitions based on provided parameters.
}
\examples{
# Example usage:
#get_temperature_summary_definitions(by_definition_list = my_definition_list, by_definition_2_list = my_definition_list_2)
}
