% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_crop_definitions.R
\name{get_values_by_sequence_or_list}
\alias{get_values_by_sequence_or_list}
\title{Get Values by Sequence or List

This function returns a sequence of values from the specified property in the \code{definitions}}
\usage{
get_values_by_sequence_or_list(definitions, property)
}
\arguments{
\item{definitions}{A list containing data or metadata properties from which to extract values.}

\item{property}{A character string giving the name of the property to retrieve from the definitions object.}
}
\value{
A vector (sequence) of values corresponding to the requested property.
}
\description{
Get Values by Sequence or List

This function returns a sequence of values from the specified property in the \code{definitions}
}
