% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_extremes_summaries.R
\name{overall_extremes_summaries}
\alias{overall_extremes_summaries}
\title{Get the Extreme Data}
\usage{
overall_extremes_summaries(
  country,
  station_id,
  summaries = c("extremes_rain", "extremes_tmin", "extremes_tmax")
)
}
\arguments{
\item{country}{A character string specifying the country code of the data.}

\item{station_id}{A character vector specifying the ID(s) of the station(s) to analyse.}

\item{summaries}{A character vector specifying the names of the summaries to produce.}
}
\value{
A data frame containing the extreme data.
}
\description{
This function identifies extreme values in a specified element (column) of a data frame. It can operate in two modes: percentile-based and threshold-based.
}
\examples{
# Generate annual temperature summaries for station 16 in Zambia
#overall_extremes_summaries(country, station_id, c("extremes_rain"), to = "annual")
}
