% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_auth_file.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{gcs_auth_file}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{epicsadata}{\code{\link[epicsadata]{gcs_auth_file}}}
}}

