% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_functions.R
\name{reformat_annual_summaries}
\alias{reformat_annual_summaries}
\title{Reformat annual summaries data}
\usage{
reformat_annual_summaries(
  data,
  station_col = NULL,
  year_col = NULL,
  start_rains_doy_col = NULL,
  start_rains_date_col = NULL,
  end_rains_doy_col = NULL,
  end_rains_date_col = NULL,
  end_season_doy_col = NULL,
  end_season_date_col = NULL,
  seasonal_rain_col = NULL,
  n_seasonal_rain_col = NULL,
  season_length_col = NULL,
  annual_rain_col = NULL,
  n_rain_col = NULL
)
}
\arguments{
\item{data}{A data frame containing the annual summaries data.}

\item{station_col}{Name of the column containing station information.}

\item{year_col}{Name of the column containing year information.}

\item{start_rains_doy_col}{Name of the column containing start of rains day of year.}

\item{start_rains_date_col}{Name of the column containing start of rains date.}

\item{end_rains_doy_col}{Name of the column containing end of rains day of year.}

\item{end_rains_date_col}{Name of the column containing end of rains date.}

\item{end_season_doy_col}{Name of the column containing end of season day of year.}

\item{end_season_date_col}{Name of the column containing end of season date.}

\item{seasonal_rain_col}{Name of the column containing seasonal rain data.}

\item{n_seasonal_rain_col}{Name of the column containing number of seasonal rain events.}

\item{season_length_col}{Name of the column containing season length.}

\item{annual_rain_col}{Name of the column containing annual rain data.}

\item{n_rain_col}{Name of the column containing number of rain events.}
}
\value{
The reformatted data frame.
}
\description{
This function reformats annual summaries data by renaming columns and converting data types.
}
