% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_functions.R
\name{reformat_crop_success}
\alias{reformat_crop_success}
\title{Reformat crop success data}
\usage{
reformat_crop_success(
  data,
  station_col = NULL,
  total_rain_col,
  plant_day_col,
  plant_length_col,
  prop_success_col
)
}
\arguments{
\item{data}{A data frame containing the crop success data.}

\item{station_col}{Name of the column containing station information.}

\item{total_rain_col}{Name of the column containing total rain data.}

\item{plant_day_col}{Name of the column containing plant day data.}

\item{plant_length_col}{Name of the column containing plant length data.}

\item{prop_success_col}{Name of the column containing proportion of success data.}
}
\value{
The reformatted data frame.
}
\description{
This function reformats crop success data by renaming columns and converting data types.
}
