% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_functions.R
\name{reformat_season_start}
\alias{reformat_season_start}
\title{Reformat season start data}
\usage{
reformat_season_start(
  data,
  station_col = NULL,
  year_col,
  plant_day_col,
  plant_day_cond_col
)
}
\arguments{
\item{data}{A data frame containing the season start data.}

\item{station_col}{Name of the column containing station information.}

\item{year_col}{Name of the column containing year information.}

\item{plant_day_col}{Name of the column containing plant day data.}

\item{plant_day_cond_col}{Name of the column containing plant day condition data.}
}
\value{
The reformatted data frame.
}
\description{
This function reformats season start data by renaming columns and calculating proportions.
}
