% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_functions.R
\name{reformat_spells_data}
\alias{reformat_spells_data}
\title{Reformat Spells data}
\usage{
reformat_spells_data(data, station_col = NULL, date_col, spells_col)
}
\arguments{
\item{data}{A data frame containing the season start data.}

\item{station_col}{Name of the column containing station information.}

\item{date_col}{Name of the column containing the date that the spell finishes.}

\item{spells_col}{Name of the column containing the spell lengths.}
}
\value{
The reformatted data frame.
}
\description{
This function reformats spells data by renaming columns.
}
