% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_functions.R
\name{reformat_temperature_summaries}
\alias{reformat_temperature_summaries}
\title{Reformat temperature summaries data}
\usage{
reformat_temperature_summaries(
  data,
  station_col = NULL,
  year_col = NULL,
  month_col = NULL,
  mean_tmin_col = NULL,
  min_tmin_col = NULL,
  max_tmin_col = NULL,
  mean_tmax_col = NULL,
  min_tmax_col = NULL,
  max_tmax_col = NULL
)
}
\arguments{
\item{data}{A data frame containing the temperature summaries data.}

\item{station_col}{Name of the column containing station information.}

\item{year_col}{Name of the column containing year information.}

\item{month_col}{Name of the column containing month information.}

\item{mean_tmin_col}{Name of the column containing mean minimum temperature data.}

\item{min_tmin_col}{Name of the column containing minimum minimum temperature data.}

\item{max_tmin_col}{Name of the column containing maximum minimum temperature data.}

\item{mean_tmax_col}{Name of the column containing mean maximum temperature data.}

\item{min_tmax_col}{Name of the column containing minimum maximum temperature data.}

\item{max_tmax_col}{Name of the column containing maximum maximum temperature data.}
}
\value{
The reformatted data frame.
}
\description{
This function reformats temperature summaries data by renaming columns and converting data types.
}
