% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_start_probabilities.R
\name{season_start_probabilities}
\alias{season_start_probabilities}
\title{Season start date probabilities}
\usage{
season_start_probabilities(country, station_id)
}
\arguments{
\item{country}{\code{character(1)} The country code of the data.}

\item{station_id}{\code{character} The id's of the stations to analyse. Either a
single value or a vector.}
}
\value{
A list containing the definitions and a data frame with probability summaries.
}
\description{
A table containing the probabilities of the season starting on or before a set of particular dates.
}
\examples{
#
#library(epicsawrap)
#library(tidyverse)
#epicsawrap::setup(dir = getwd())
#epicsawrap::gcs_auth_file(file = "C:/Users/lclem/Downloads/e-picsa-e630400792e7.json")
#season_start_probabilities(country = "zm", station_id = "16")
}
