% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup directories}
\usage{
setup(dir, countries = c("mw", "zm", "zm_test", "zw"))
}
\arguments{
\item{dir}{\code{character(1)} The path to set as the working directory}

\item{countries}{\code{character(1)} The set of countries to create directories
for.}
}
\value{
TODO
}
\description{
Setup directories
}
\examples{
#TODO
}
