% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_metadata.R
\name{station_metadata}
\alias{station_metadata}
\title{Get Station Metadata}
\usage{
station_metadata(country)
}
\arguments{
\item{country}{\code{character(1)} The country code of the data.}
}
\value{
TODO
}
\description{
Get Station Metadata
}
\examples{
# TODO
}
