% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_metadata_definitions.R
\name{station_metadata_definitions}
\alias{station_metadata_definitions}
\title{Get Processed Station Metadata}
\usage{
station_metadata_definitions(
  country,
  station_id,
  format = c("wide", "long", "nested", "list")
)
}
\arguments{
\item{country}{A character vector specifying the country code for which station metadata should be retrieved and processed. Options are defined in \code{get_bucket_name()} (e.g., \code{"zm"}, \code{"mw"}).}

\item{station_id}{A character vector specifying the station ID(s) for the given country.}

\item{format}{A character vector indicating the desired format of the processed data. It can be "wide", "long", "nested", or "list".}
}
\value{
Depending on the specified format, the function returns the processed station metadata in either wide, long, nested, or list format.
}
\description{
This function retrieves and processes station metadata for the specified country and format. The station metadata includes information about station IDs and their associated definitions.
}
\examples{
# Retrieve and process station metadata for country "zm" in wide format
#station_metadata_definitions(country = "zm", format = "wide")

# Retrieve and process station metadata for countries "zm" and "mw" in long format
#station_metadata_definitions(country = c("zm", "mw"), format = "long")

# Retrieve and process station metadata for country "zm" in nested format
#station_metadata_definitions(country = "zm", format = "nested")
}
