% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_temperature_summaries.R
\name{total_temperature_summaries}
\alias{total_temperature_summaries}
\title{Annual and Monthly Temperature Summaries}
\usage{
total_temperature_summaries(
  country,
  station_id,
  summaries = c("mean_tmin", "mean_tmax"),
  to = c("annual", "monthly")
)
}
\arguments{
\item{country}{\code{character(1)} The country code of the data.}

\item{station_id}{\code{character} The id's of the stations to analyse. Either a
single value or a vector.}

\item{summaries}{\code{character} The names of the summaries to produce.}

\item{to}{\code{character} Whether this is for annual or monthly summaries}
}
\value{
A data frame with either annual or monthly summaries.
}
\description{
Annual and Monthly Temperature Summaries
}
\examples{
# total_temperature_summaries(country = "zm", station_id = "16") # made a fake "16" json definitions data
# because it contains temperature data. 
}
