% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all_definitions.R
\name{update_all_definitions}
\alias{update_all_definitions}
\title{Update All Definitions}
\usage{
update_all_definitions(
  data_frame,
  data_names,
  summary_data_frame,
  summary_data_names,
  definitions,
  data_book
)
}
\arguments{
\item{data_frame}{A data frame containing the main dataset whose definitions are to be updated.}

\item{data_names}{A list of column names that are needed from  \code{data_frame}.}

\item{summary_data_frame}{A data frame containing summary data corresponding to the main dataset.}

\item{summary_data_names}{A list of column names that are needed from \code{summary_data_frame}.}

\item{definitions}{A list containing metadata or variable definitions to be updated.}

\item{data_book}{The data book object where the data object is stored.}
}
\value{
The updated data for all of the definition files.
}
\description{
This function updates the data for all the definitions
}
