% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/update_total_temperature_summaries_from_definition.R
\name{update_annual_temperature_summaries_from_definition}
\alias{update_annual_temperature_summaries_from_definition}
\title{Update Annual Temperature Summaries from Definitions}
\usage{
update_annual_temperature_summaries_from_definition(
  country,
  station_id,
  daily_data
)
}
\arguments{
\item{country}{Character. The name of the country for which definitions and observations are retrieved.}

\item{station_id}{Character. The station ID(s) for which data is retrieved. Can be \code{NULL} if \code{definition_id} is specified. Defaults to \code{NULL}.}

\item{daily_data}{Data frame. Daily temperature data.}
}
\value{
A data frame containing annual temperature summaries.
}
\description{
A wrapper function to generate only annual temperature summaries based on definitions.
}
