% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_crop_definitions.R
\name{update_crops_definitions}
\alias{update_crops_definitions}
\title{Update Crops Definition}
\usage{
update_crops_definitions(
  data_frame,
  data_names,
  seasonal_data_frame,
  seasonal_data_names,
  definitions,
  data_book
)
}
\arguments{
\item{data_frame}{A data frame containing the columns specified in \code{data_names}.}

\item{data_names}{A list of column names that are needed from the dataframe.}

\item{seasonal_data_frame}{A data frame containing the columns specified in \code{seasonal_data_names}.}

\item{seasonal_data_names}{A list of column names that are needed from the \code{seasonal_data_frame}.}

\item{definitions}{A list containing definitions to be read in.}

\item{data_book}{The data book object where the data object is stored.}
}
\value{
A data frame in the data book at the year (and station) level containing crops definitions data.
}
\description{
Update Crops Definition
}
