% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_daily_data.R
\name{update_daily_data}
\alias{update_daily_data}
\title{Get the Daily Data}
\usage{
update_daily_data(country, station_id)
}
\arguments{
\item{country}{A character vector specifying the country or countries from which to update the data. Options are defined in \code{get_bucket_name()} (e.g., \code{"zm"}, \code{"mw"}).}

\item{station_id}{A character string specifying the ID of the station for which to update the daily data.}
}
\value{
This function does not return any value explicitly. It gets the daily data for the specified station in the specified country.
}
\description{
This function updates the daily data for a specific station in the specified country. It retrieves the data from Google Cloud Storage using the \code{get_data} function.
}
\details{
The \code{update_daily_data} function is used to update the daily data for a specific station in the specified country. It internally calls the \code{get_data} function to retrieve the data from Google Cloud Storage.
The \code{country} argument is a character vector that allows specifying one or more countries from which to update the data. The data will be updated for Mozambique (\code{"mz"}) and Zambia (\code{"zm"}). You can modify this argument to update data for different countries.
The \code{station_id} argument is a character string that specifies the ID of the station for which to update the daily data. The function will construct the filename by concatenating the \code{"data/"} directory, the \code{station_id}, and the \code{file} extension \code{".rds"}. The filename will be passed to the \code{get_data} function to retrieve the data.
The function uses the invisible function to suppress the output of the \code{get_data} function, ensuring that the data retrieval process is not visible in the console.
}
\examples{
# todo
}
