% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_end_season.R
\name{update_end_season}
\alias{update_end_season}
\title{Update Start Rains}
\usage{
update_end_season(data_frame, data_names, definitions, data_book)
}
\arguments{
\item{data_frame}{A data frame containing the columns specified in \code{data_names}.}

\item{data_names}{A list of column names that are needed from the \code{data_frame}.}

\item{definitions}{A list containing definitions to be read in.}

\item{data_book}{The data book object where the data object is stored.}
}
\value{
A data frame in the data book at the year (and station) level containing end of season data.
}
\description{
Update Start Rains
}
