% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_get_extremes.R
\name{update_get_extremes}
\alias{update_get_extremes}
\title{Update Get Extremes}
\usage{
update_get_extremes(
  data_frame,
  data_names,
  definitions,
  data_book,
  element = c("extreme_rain", "extreme_tmin", "extreme_tmax")
)
}
\arguments{
\item{data_frame}{A data frame containing the columns specified in \code{data_names}.}

\item{data_names}{A list of column names that are needed from the \code{data_frame}.}

\item{definitions}{A list containing definitions to be read in.}

\item{data_book}{The data book object where the data object is stored.}

\item{element}{Determines the specific variant of the definitions to use.}
}
\value{
A data frame in the data book at the year (and station) level containing extremes data.
}
\description{
Update Get Extremes
}
