% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata.R
\name{update_metadata}
\alias{update_metadata}
\title{Update Station Metadata}
\usage{
update_metadata(country)
}
\arguments{
\item{country}{A character vector specifying the country or countries from which to update the metadata. Options are defined in \code{get_bucket_name()} (e.g., \code{"zm"}, \code{"mw"}).}
}
\value{
This function updates the metadata for the specified station in the specified country.
}
\description{
Update Station Metadata
}
\examples{
#todo
}
