% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata_definition_id.R
\name{update_metadata_definition_id}
\alias{update_metadata_definition_id}
\title{Update Metadata Definition ID}
\usage{
update_metadata_definition_id(
  country,
  station_id,
  definition_id,
  overwrite = FALSE,
  add_climsoft = FALSE,
  elementfiltercolumn = "elementName",
  elements = c("Temp  Daily Max", "Temp  Daily Min", "Precip  Daily")
)
}
\arguments{
\item{country}{A character string representing the country code.}

\item{station_id}{A character string representing the station ID.}

\item{definition_id}{A character string representing the new definition ID to be added.}

\item{overwrite}{A logical value indicating whether to overwrite the existing definitions ID.}

\item{add_climsoft}{A logical value indicating whether to add in climsoft details. This will add in the values from \code{elementfiltercolumn} and \code{elements}.}

\item{elementfiltercolumn}{Name of the column to filter by elements, default is 'elementName'.}

\item{elements}{Vector of element IDs to filter the data.
If \code{TRUE}, the existing definitions ID will be overwritten. If \code{FALSE}, the new
definition ID will be appended. Default is \code{FALSE}.}
}
\value{
None. The function updates the metadata in the specified cloud storage bucket.
}
\description{
This function updates the definitions ID in the station metadata for a given country and station ID.
It can either overwrite the existing definitions ID or append a new one.
}
