% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_monthly_temperature.R
\name{update_monthly_temperature}
\alias{update_monthly_temperature}
\title{Update Monthly Temperature Data}
\usage{
update_monthly_temperature(data_frame, data_names, definitions, data_book)
}
\arguments{
\item{data_frame}{A data frame containing the columns specified in \code{data_names}.}

\item{data_names}{A list of column names that are needed from the \code{data_frame}.}

\item{definitions}{A list containing definitions to be read in.}

\item{data_book}{The data book object where the data object is stored.}
}
\value{
A data frame in the data book at the month (and station) level containing monthly temperature data
}
\description{
Update Monthly Temperature Data
}
