% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_summaries_functions.R
\name{update_season_start_probabilities_from_definition}
\alias{update_season_start_probabilities_from_definition}
\title{Update Season Start Probabilities from Definitions}
\usage{
update_season_start_probabilities_from_definition(
  country,
  station_id = NULL,
  definition_id = NULL,
  daily_data = NULL,
  start_rains_data = NULL
)
}
\arguments{
\item{country}{Character. The name of the country for which definitions and observations are retrieved.}

\item{station_id}{Character. The station ID(s) for which data is retrieved. Can be \code{NULL} if \code{definition_id} is specified. Defaults to \code{NULL}.}

\item{definition_id}{Character. The ID of the definitions to use for generating summaries. Only used if \code{station_id} is \code{NULL}. Defaults to \code{NULL}.}

\item{daily_data}{Data frame. (Optional) Daily rainfall data, required if \code{start_rains_data} is not provided. Defaults to \code{NULL}.}

\item{start_rains_data}{Data frame. (Optional) Precomputed start-of-rains data. If not provided, it is generated using \code{daily_data}. Defaults to \code{NULL}.}
}
\value{
A data frame containing the probabilities of the season starting on specified days, based on the definitions provided.
}
\description{
This function calculates the probabilities of the season starting on specified days based on the most recent definitions
for a given country and station. It uses rainfall data or start-of-rains summaries to compute the probabilities.
}
\details{
\itemize{
\item The function retrieves definition data to identify the days of interest for calculating the start-of-season probabilities.
\item If \code{start_rains_data} is not provided, it computes the start-of-rains summaries using \code{daily_data}.
\item The function calculates probabilities for the specified days using the \code{probability_season_start} function in \code{rpicsa}
}
}
