% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_summary_functions.R
\name{annual_rainfall_summaries}
\alias{annual_rainfall_summaries}
\title{Annual Rainfall Summaries}
\usage{
annual_rainfall_summaries(
  country,
  station_id,
  call = c("climsoft", "googlebuckets"),
  summaries = c("annual_rain", "start_rains", "end_rains", "end_season", "seasonal_rain",
    "seasonal_length"),
  override = FALSE
)
}
\arguments{
\item{country}{\code{character(1)} The country code of the data.}

\item{station_id}{\code{character} The id's of the stations to analyse. Either a
single value or a vector.}

\item{call}{A character vector specifying where to call the raw data from if calling raw data.}

\item{summaries}{\code{character} The names of the summaries to produce.}

\item{override}{A logical argument default \code{FALSE} indicating whether to calculate the summaries still, even if they are stored already in the bucket.}
}
\value{
A data frame with yearly summaries.
}
\description{
A table containing all the annual rainfall summaries for PICSA
e.g. start of rain, total rainfall, number of rain days, end of season.
One row per year/station and one column per summary.
}
\examples{
#annual_rainfall_summaries(country = "zm", station_id = "01122", summaries = "annual_rain")
#annual_rainfall_summaries(country = "zm", station_id = "16", summaries = c("start_rains", "end_rains", "annual_rain", "seasonal_rain")) #, "end_season"))
}
