% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_temperature_summaries.R
\name{annual_temperature_summaries}
\alias{annual_temperature_summaries}
\title{Annual Temperature Summaries}
\usage{
annual_temperature_summaries(
  country,
  station_id,
  summaries = c("mean_tmin", "mean_tmax")
)
}
\arguments{
\item{country}{\code{character(1)} The country code of the data.}

\item{station_id}{\code{character} The id's of the stations to analyse. Either a
single value or a vector.}

\item{summaries}{\code{character} The names of the summaries to produce.}
}
\value{
A data frame with yearly summaries.
}
\description{
Annual Temperature Summaries
}
\examples{
# annual_temperature_summaries(country = "zm", station_id = "01122")
}
