% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_total_temperature_summaries.R
\name{build_total_temperature_summaries}
\alias{build_total_temperature_summaries}
\title{Calculate total temperature summaries}
\usage{
build_total_temperature_summaries(
  year = data_book$get_climatic_column_name(data_name, "year"),
  month = data_book$get_climatic_column_name(data_name, "month"),
  data_by_year,
  data_by_year_month,
  min_tmin_column,
  mean_tmin_column,
  max_tmin_column,
  min_tmax_column,
  mean_tmax_column,
  max_tmax_column
)
}
\arguments{
\item{year}{Character vector specifying the year.}

\item{month}{Character vector specifying the month.}

\item{data_by_year}{A list of temperature summaries by definition (e.g., year).}

\item{data_by_year_month}{An optional second list of temperature summaries by definition (e.g., year and month).}

\item{min_tmin_column}{The name of the minimum of minimum temperature column in the data.}

\item{mean_tmin_column}{The name of the mean of minimum temperature column in the data.}

\item{max_tmin_column}{The name of the maximum of minimum temperature column in the data.}

\item{min_tmax_column}{The name of the minimum of maximum temperature column in the data.}

\item{mean_tmax_column}{The name of the mean of maximum temperature column in the data.}

\item{max_tmax_column}{The name of the maximum of maximum temperature column in the data.}
}
\value{
A list containing total temperature summaries.
}
\description{
Calculates total temperature summaries based on provided parameters.
}
\examples{
# Example usage:
#total_temperature_summaries(tmin = "tmin", tmax = "tmax", year = "year", month = "month", 
#                             data_by_year = my_definition_list, data_by_year_month = my_definition_list_2)
}
