% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{check_and_rename_variables}
\alias{check_and_rename_variables}
\title{Check and Rename Variables in a Dataset}
\usage{
check_and_rename_variables(data, data_names)
}
\arguments{
\item{data}{A data frame or data table containing the dataset.}

\item{data_names}{A named character vector where the names are the standard variable names,
and the values are the corresponding variable names in the dataset.}
}
\value{
A data frame with variable names renamed according to \code{data_names}.
}
\description{
This function checks the variable names in a dataset and renames them based on
the provided \code{data_names} vector.
}
