% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_success_probabilities.R
\name{crop_success_probabilities}
\alias{crop_success_probabilities}
\title{Probability Crop Tables}
\usage{
crop_success_probabilities(
  country,
  station_id,
  planting_dates = NULL,
  water_requirements = NULL,
  crop_length = NULL,
  start_before_season = NULL
)
}
\arguments{
\item{country}{\code{character(1)} The country code of the data.}

\item{station_id}{\code{character} The id's of the stations to analyse. Either a
single value or a vector.}

\item{plant_days}{\code{character} Vector containing a list of seasonal total rainfall requirements.}

\item{plant_lengths}{\code{character} Vector containing a list of seasonal total rainfall requirements.}

\item{start_check}{\code{logical} A logical value indicating whether to check the start day condition (default is \code{TRUE}).}

\item{season_data}{The data frame containing the seasonal data.}
}
\value{
A list containing the definitions and a data frame with probability summaries.
}
\description{
The probabilities of crop success for given planting maturity lengths, seasonal total rainfall requirements, and planting dates.
}
\examples{
#
#library(epicsawrap)
#library(tidyverse)
#epicsawrap::setup(dir = getwd())
#epicsawrap::gcs_auth_file(file = "C:/Users/lclem/Downloads/e-picsa-e630400792e7.json")
#crop_success_probabilities(country = "zm", station_id = "16")

# or some can be defined in the dialog
#crop_success_probabilities(country = "zm", station_id = "16", water_requirements = c(100, 300, 800))
}
