% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_success_probabilities.R
\name{crop_success_probabilities}
\alias{crop_success_probabilities}
\title{Probability Crop Tables}
\usage{
crop_success_probabilities(country, station_id)
}
\arguments{
\item{country}{\code{character(1)} The country code of the data.}

\item{station_id}{\code{character} The id's of the stations to analyse. Either a
single value or a vector.}
}
\value{
A list containing the definitions and a data frame with probability summaries.
}
\description{
The probabilities of crop success for given planting maturity lengths, seasonal total rainfall requirements, and planting dates.
}
\examples{
#
#library(epicsawrap)
#library(tidyverse)
#epicsawrap::setup(dir = getwd())
#epicsawrap::gcs_auth_file(file = "C:/Users/lclem/Downloads/e-picsa-e630400792e7.json")
#crop_success_probabilities(country = "zm", station_id = "16")
}
