% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{definitions}
\alias{definitions}
\title{Definitions}
\usage{
definitions(
  country,
  station_id = NULL,
  definitions_id = NULL,
  summaries,
  file = NULL
)
}
\arguments{
\item{country}{\code{character(1)} The country code of the data.}

\item{station_id}{\code{character(1)} The definitions code in the data.}

\item{definitions_id}{\code{character(1)} The definitions code in the data.}

\item{summaries}{\code{character} Vector of summaries to display}

\item{file}{Default \code{NULL} meaning that the most recent definitions file will be found and imported. Otherwise specify as a string the file to import. In format: "STATIONNAME.TIMESTAMP" e.g. "1.20240311152831"}
}
\value{
TODO
}
\description{
Definitions
}
\examples{
# e.g. definitions("zm", "16", "annual_rain")
# error: definitions("zm", "1", c("annual_rain", "hi", "end_season"))
}
