% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annual_rain_definitions.R,
%   R/get_total_rain_counts.R
\name{get_total_rain_counts}
\alias{get_total_rain_counts}
\title{Get annual rain definitions}
\usage{
get_total_rain_counts(
  data_name,
  data_by_year = NULL,
  rain_name = data_book$get_climatic_column_name(data_name = "ghana", col_name = "rain")
)

get_total_rain_counts(
  data_name,
  data_by_year = NULL,
  rain_name = data_book$get_climatic_column_name(data_name = "ghana", col_name = "rain")
)
}
\arguments{
\item{data_name}{Character string specifying the name of the data.}

\item{data_by_year}{List containing data by year.}

\item{rain_name}{Character string specifying the name of the rainfall data.}
}
\value{
A list representation of annual rain definitions.

A list containing definitions of total rain counts for annual and seasonal periods.
}
\description{
Retrieves annual rain definitions.

This function retrieves the definition for the total rainfall as well as the
total number of rainy days for both annual and seasonal periods.
}
\examples{
# Example usage:
#get_total_rain_counts(annual_rain, ghana_defs)

#data <- your_data
#get_total_rain_counts(data_name = "ghana", data_by_year = data, rain_name = "rain")
}
