% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{monthly_temperature_summaries}
\alias{monthly_temperature_summaries}
\title{Monthly Temperature Summaries}
\usage{
monthly_temperature_summaries(
  country,
  station_id,
  call = c("climsoft", "googlebuckets"),
  summaries = c("mean_tmin", "mean_tmax", "min_tmin", "min_tmax", "max_tmin", "max_tmax"),
  override = FALSE
)
}
\arguments{
\item{country}{\code{character(1)} The country code of the data.}

\item{station_id}{\code{character} The id's of the stations to analyse. Either a
single value or a vector.}

\item{call}{A character vector specifying where to call the raw data from if calling raw data.}

\item{summaries}{\code{character} The names of the summaries to produce.}

\item{override}{A logical argument default \code{FALSE} indicating whether to calculate the summaries still, even if they are stored already in the bucket.}
}
\value{
A data frame with monthly summaries.
}
\description{
Monthly Temperature Summaries
}
