% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_summary_functions.R
\name{overall_extremes_summaries}
\alias{overall_extremes_summaries}
\title{Generate summary statistics for extreme weather events.}
\usage{
overall_extremes_summaries(daily, data_names, definitions, summaries)
}
\arguments{
\item{daily}{A dataframe containing daily weather data.}

\item{data_names}{A list containing the data names in the \code{daily} data.}

\item{definitions}{A list containing definitions of extreme events.}

\item{summaries}{Name of the summary to be generated.}
}
\value{
A dataframe containing summary statistics for extreme events.
}
\description{
This function generates summary statistics for extreme weather events based on given definitions.
}
\examples{
# Example usage:
# Generate summary statistics for extreme rain events
#rain_summary <- overall_extremes_summaries(daily_data, definitions_list, "extremes_rain")
}
