% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_summary_functions.R
\name{season_start_probabilities}
\alias{season_start_probabilities}
\title{Season start date probabilities}
\usage{
season_start_probabilities(
  country,
  station_id,
  call = c("climsoft", "googlebuckets"),
  start_dates = NULL,
  override = FALSE
)
}
\arguments{
\item{country}{\code{character(1)} The country code of the data.}

\item{station_id}{\code{character} The id's of the stations to analyse. Either a
single value or a vector.}

\item{call}{A character vector specifying where to call the raw data from if calling raw data.}

\item{start_dates}{\code{numeric} A vector of start dates (in doy format) to calculate the probabilities of the season starting on or before.}

\item{override}{A logical argument default \code{FALSE} indicating whether to calculate the summaries still, even if they are stored already in the bucket.}
}
\value{
A list containing the definitions and a data frame with probability summaries.
}
\description{
A table containing the probabilities of the season starting on or before a set of particular dates.
}
\examples{
#
#library(epicsawrap)
#library(tidyverse)
#epicsawrap::setup(dir = getwd())
#epicsawrap::gcs_auth_file(file = "C:/Users/lclem/Downloads/e-picsa-e630400792e7.json")
#season_start_probabilities(country = "zm", station_id = "16")
# or you can manually define
#season_start_probabilities(country = "zm", station_id = "16", start_dates = c(10, 20, 100))
}
