% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_temperature_summaries.R
\name{total_temperature_summaries}
\alias{total_temperature_summaries}
\title{Generate Annual and Monthly Temperature Summaries}
\usage{
total_temperature_summaries(
  country,
  station_id,
  summaries = c("mean_tmin", "mean_tmax"),
  to = c("annual", "monthly")
)
}
\arguments{
\item{country}{A character string specifying the country code of the data.}

\item{station_id}{A character vector specifying the ID(s) of the station(s) to analyze.}

\item{summaries}{A character vector specifying the names of the summaries to produce.}

\item{to}{A character string indicating whether the summaries should be generated for "annual" or "monthly" data.}
}
\value{
A data frame containing the requested temperature summaries.
}
\description{
This function calculates annual or monthly temperature summaries for specified stations in a given country.
}
\examples{
# Generate annual temperature summaries for station 16 in Zambia
#total_temperature_summaries(country = "zm", station_id = "16", to = "annual")
}
