% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_summary_functions.R
\name{total_temperature_summaries}
\alias{total_temperature_summaries}
\title{Generate Annual and Monthly Temperature Summaries}
\usage{
total_temperature_summaries(
  country,
  station_id,
  call = c("climsoft", "googlebuckets"),
  summaries = c("mean_tmin", "mean_tmax", "min_tmin", "min_tmax", "max_tmin", "max_tmax"),
  to = c("annual", "monthly"),
  override = FALSE
)
}
\arguments{
\item{country}{A character string specifying the country code of the data.}

\item{station_id}{A character vector specifying the ID(s) of the station(s) to analyse.}

\item{call}{A character vector specifying where to call the raw data from if calling raw data.}

\item{summaries}{A character vector specifying the names of the summaries to produce.}

\item{to}{A character string indicating whether the summaries should be generated for "annual" or "monthly" data.}

\item{override}{A logical argument default \code{FALSE} indicating whether to calculate the summaries still, even if they are stored already in the bucket.}
}
\value{
A data frame containing the requested temperature summaries.
}
\description{
This function calculates annual or monthly temperature summaries for specified stations in a given country.
}
\examples{
# Generate annual temperature summaries for station 16 in Zambia
#total_temperature_summaries(country = "zm", station_id = "1", summaries = c("mean_tmin", "mean_tmax", "min_tmin", "max_tmax"), to = "annual")
}
