% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_end_rains.R
\name{update_end_rains}
\alias{update_end_rains}
\title{Update End of Rains Definition}
\usage{
update_end_rains(data_frame, data_names, definitions, data_book)
}
\arguments{
\item{data_frame}{A data frame containing the columns specified in \code{data_names}.}

\item{data_names}{A list of column names that are needed from the dataframe.}

\item{definitions}{A list containing definitions to be read in.}

\item{data_book}{The data book object where the data object is stored.}
}
\value{
A data frame in the data book at the year (and station) level containing end of rains data.
}
\description{
Runs the \code{end_rains} function from the \code{rpicsa} package by reading in the parameter values from the
definitions
}