% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_summaries_from_definition.R
\name{update_rainfall_summaries_from_definition}
\alias{update_rainfall_summaries_from_definition}
\title{Update Rainfall Summaries from Definitions}
\usage{
update_rainfall_summaries_from_definition(
  country = "zm_workshops",
  station_id = "Lundazi Met",
  daily_data,
  summaries = c("annual_rain", "start_rains", "end_rains", "end_season", "seasonal_rain",
    "seasonal_length")
)
}
\arguments{
\item{country}{Character. The name of the country for which definitions and observations are retrieved. Defaults to "zm_workshops".}

\item{station_id}{Character. The station ID for which data is retrieved. Defaults to "Lundazi Met".}

\item{daily_data}{Data frame. The daily data to update.}

\item{summaries}{\code{character} The names of the summaries to produce.}
}
\value{
A data frame containing summarized rainfall data for the specified station and definitions.
}
\description{
This function generates summary data based on the most recent definitions and observed rainfall data
for a specified country and weather station. It calculates various metrics, such as start and end
of rains, seasonal length, seasonal rainfall, and annual rainfall, depending on the definitions provided.
}
\examples{
#update_rainfall_summaries_from_definition(country = "zm_workshops", station_id = "Lundazi Met")
}
