% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_seasonal_rain.R
\name{update_seasonal_rain}
\alias{update_seasonal_rain}
\title{Update Seasonal Rain Data}
\usage{
update_seasonal_rain(
  data_frame,
  data_names,
  summary_data_frame,
  summary_data_names,
  definitions,
  data_book
)
}
\arguments{
\item{data_frame}{A data frame containing the columns specified in \code{data_names}.}

\item{data_names}{A list of column names that are needed from the \code{data_frame}.}

\item{summary_data_frame}{A summary data frame containing the columns specified in \code{summary_data_names}.}

\item{summary_data_names}{A list of column names that are needed from the \code{summary_data_frame}.}

\item{definitions}{A list containing definitions to be read in.}

\item{data_book}{The data book object where the data object is stored.}
}
\value{
A data frame in the data book at the year (and station) level containing seasonal rainfall data.
}
\description{
Update Seasonal Rain Data
}
