% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_spells.R
\name{update_spells}
\alias{update_spells}
\title{Update Spells}
\usage{
update_spells(
  data_frame,
  data_names,
  seasonal_data_frame = NULL,
  seasonal_data_names = NULL,
  definitions,
  element = c("rain", "tmin", "tmax"),
  data_book
)
}
\arguments{
\item{data_frame}{A data frame (or name) containing daily data.}

\item{data_names}{A list with element names for daily data:
\code{date}, \code{year}, \code{station}, \code{doy}, \code{rain}.}

\item{seasonal_data_frame}{A data frame (or name) containing seasonal windows.}

\item{seasonal_data_names}{A list with element names for seasonal data:
\code{start_day}, \code{end_day}. If these are not given, then values from
the definitions file are used.}

\item{definitions}{A list with (optional) elements
\code{$annual_summaries$longest_rain_spell} or
\code{$annual_summaries$longest_tmin_spell} or
\code{$annual_summaries$longest_tmax_spell}.}

\item{element}{Which element to apply the spells to. One of \code{"rain"}, \code{"tmin"}, or \code{"tmax"}.}

\item{data_book}{The data book object forwarded to
\code{rpicsa::crops_definitions()}.}
}
\value{
A data frame in the data book at the year (and station) level containing spells data.
}
\description{
Update Spells
}
