% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_spells.R
\name{update_spells}
\alias{update_spells}
\title{Update Spells}
\usage{
update_spells(
  data_frame,
  data_names,
  summary_data_frame,
  definitions,
  data_book
)
}
\arguments{
\item{data_frame}{A data frame containing the columns specified in \code{data_names}.}

\item{data_names}{A list of column names that are needed from the \code{data_frame}.}

\item{definitions}{A list containing definitions to be read in.}

\item{data_book}{The data book object where the data object is stored.}

\item{summary_data}{A data frame of the computed summaries.}
}
\value{
A data frame in the data book at the year (and station) level containing spells data.
}
\description{
Update Spells
}
