% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_start_rains.R
\name{update_start_rains}
\alias{update_start_rains}
\title{Update Start Rains}
\usage{
update_start_rains(data_frame, data_names, definitions, data_book)
}
\arguments{
\item{data_frame}{A data frame containing the columns specified in \code{data_names}.}

\item{data_names}{A list of column names that are needed from the \code{data_frame}.}

\item{definitions}{A list containing definitions to be read in.}

\item{data_book}{The data book object where the data object is stored.}
}
\description{
Update Start Rains
}
\examples{
library(rpicsa)
library(databook)

# 1. Let's set up our data book
data_book <- DataBook$new()

# 2. Importing in some data for testing (this is stored in the rpicsa package)
data(daily_niger)
data_book$import_data(list(daily_niger = daily_niger))

# 3. Read in our definitions data
definitions <- jsonlite::fromJSON("C:/Users/HP/Downloads/test_json_1.json")

# 4. Put in "data_names" the names of all the variables we're going to use from the daily_niger data. 
# Looking at our rpicsa::annual_rain, this can be
# station, year, and rain
data_names <- list(date = "date", 
                   rain = "rain",
                   year = "year", 
                   doy = "doy",
                   station = "station_name")


\dontrun{
  update_start_rains(data_frame = "daily_niger", 
                     data_names = data_names, 
                     definitions = definitions, 
                     data_book = data_book)
}

}
