% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_summaries_functions.R
\name{update_total_temperature_summaries_from_definition}
\alias{update_total_temperature_summaries_from_definition}
\title{Update Total Temperature Summaries from Definitions}
\usage{
update_total_temperature_summaries_from_definition(
  country = "zm_workshops",
  station_id = NULL,
  definition_id = NULL,
  daily_data,
  to = c("annual", "monthly")
)
}
\arguments{
\item{country}{Character. The name of the country for which definitions and observations are retrieved. Defaults to "zm_workshops".}

\item{station_id}{Character. The station ID(s) for which data is retrieved. Can be \code{NULL} if \code{definition_id} is specified. Defaults to \code{NULL}.}

\item{definition_id}{Character. The ID of the definitions to use for generating summaries. Only used if \code{station_id} is \code{NULL}. Defaults to \code{NULL}.}

\item{daily_data}{Data frame. Daily temperature data, including columns for station, date, year, tmin, and tmax.}

\item{to}{Character. Specifies whether the summaries should be generated for "annual" or "monthly" periods. Defaults to "annual".}
}
\value{
A data frame containing temperature summaries based on the specified definitions and period (annual or monthly).
}
\description{
This function computes temperature summaries (e.g., minimum, maximum, mean) based on the most recent definitions
for a specified country and station. It supports generating either annual or monthly summaries.
}
